requirements.txt
pathway
llm_app
unstructured==0.10.15
unstructured[pdf]==0.10.15
tiktoken
requests
datetime
streamlit
python-dotenv